Moved the playlist image saving to another converter
